derive: move Syntax into parser module
